+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # 1.0.8
 
 - added the `KafkaLoad` stage - a stage which will write a dataset to a designated `topic` in Apache Kafka.
+- added the `EXPERIMENTAL` `KafkaExtract` stage - a stage which will read from a designated `topic` in Apache Kafka and produce a dataset.
 
 # 1.0.7